+++ src/common/username.c
@@ -40,8 +40,8 @@ get_user_name(char **errstr)
 	pw = getpwuid(user_id);
 	if (!pw)
 	{
-		*errstr = psprintf(_("could not look up effective user ID %ld: %s"),
-						   (long) user_id,
+		*errstr = psprintf(_("could not look up effective user ID %d: %s"),
+						   user_id,
 						   errno ? strerror(errno) : _("user does not exist"));
 		return NULL;
 	}
